requirements.txt
cycler==0.10.0
boto==2.39.0
cffi==1.14.0
attrs==19.2.0
httplib2==0.9.2
Babel==2.6.0
cloudpickle==0.8.0
coverage==3.6
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.0.7
djangorestframework==3.11.0
django-heroku==0.3.1
docutils==0.14
spacy==2.2.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm
gunicorn==19.10.0
hyperlink==19.0.0
importlib-metadata==1.5.0
jsonschema==3.0.1
nltk==3.4
pickleshare==0.7.5
python-dotenv==0.14.0
psutil==5.6.1
psycopg2==2.8.5
pycodestyle==2.5.0
pycorenlp==0.3.0
python-dateutil==2.8.0
scikit-learn==0.20.3
Scrapy==1.8.0
scrapy-djangoitem==1.1.1
srsly==1.0.1
six==1.15.0
tokenizer==1.4.1
Twisted==20.3.0
urllib3==1.24.1
virtualenv==16.7.5
virtualenv-clone==0.5.3
whitenoise==5.1.0

